remove warning

+++ SmartHome/FMController.m
@@ -167,7 +167,7 @@
         cell.userInteractionEnabled = NO;
     }else{
         TVChannel *channel = self.allFavouriteChannels[indexPath.row];
-        cell.chanelNum.text = [NSString stringWithFormat:@"%ld",channel.channel_id];
+        cell.chanelNum.text = [NSString stringWithFormat:@"%ld",(long)channel.channel_id];
         cell.channelName.text = [NSString stringWithFormat:@"%@",channel.channel_name];
         [cell useLongPressGesture];
     }
